requirements.txt
python-dotenv==1.0.1
streamlit==1.43.0
google-genai==1.4.0